Update printer-installer.source.py
Used 2to3 framework to convert this script to Python 3. Great script, but no longer works for Monterey 12.3 and onward as they finally pulled Python 2 from being shipped.

Also needed to add a *.decode('UTF-8')* on line 225 as it was still concatenating as bytes which doesn't work in Python 3.

Only known issue is the script errors if the user clicks cancel. Working to fix this as time allows, but at least it's now working in Python 3.